+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 2.2.5  (2018-XX-XX)
 ### Bugs Fixed
 
 - Issue #479: `File.Move` on a file opened with `FileShare.Delete` succeeds but throws `IOException`.  (Thx oguimbal)
+- Issue #484: `DirectoryNotFoundException` with `MoveOptions.DelayUntilReboot`  (Thx 0x084E)
 
 
 Version 2.2.4  (2018-07-12)